+++ examples/RSC/app/_storage/layout.storage.server.ts
@@ -1,6 +1,6 @@
+import { LayoutClientContext } from "@/app/_storage/layout.context.client"
+import { layoutStorage } from "@/app/_storage/layout.storage"
 import { serverify } from "@yakocloud/state-vocab/server"
-import { layoutStorage } from "@/context/layout.storage";
-import { LayoutClientContext } from "@/context/layout.context.client";
 
 export const layoutServerStorage = serverify(layoutStorage, {
   clientContext: LayoutClientContext,