+++ examples/RSC/app/_storage/page.storage.client.ts
@@ -1,0 +1,6 @@
+"use client"
+
+import { pageStorage } from "@/app/_storage/page.storage"
+import { clientify } from "@yakocloud/state-vocab/client"
+
+export const pageClientStorage = clientify(pageStorage)