+++ examples/RSC/app/_storage/page.storage.server.ts
@@ -1,0 +1,4 @@
+import { serverify } from "@yakocloud/state-vocab/server"
+import { pageStorage } from "@/app/_storage/page.storage"
+
+export const pageServerStorage = serverify(pageStorage)